fix: fixed middleware
