--- src/auth-service/middleware/root_middleware.go
@@ -1,17 +1,0 @@
-package middleware
-
-type RootMiddleware struct {
-	TransactionMiddleware *TransactionMiddleware
-	AuthMiddleware        *AuthMiddleware
-}
-
-func NewRootMiddleware(
-	transactionMiddleware *TransactionMiddleware,
-	authMiddleware *AuthMiddleware,
-) *RootMiddleware {
-	rootMiddleware := &RootMiddleware{
-		TransactionMiddleware: transactionMiddleware,
-		AuthMiddleware:        authMiddleware,
-	}
-	return rootMiddleware
-}
